backend/requirements.txt
ipykernel
ipywidgets
transformers
torch
huggingface_hub[cli]
sentence-transformers
lancedb
pandas
python-dotenv
fastapi[standard]
openai
